# go/model_position.go
/*
 * ExCommerce
 *
 * ExCommerce is an example commerce system.
 *
 * API version: beta
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

// Position - A position in a cart.
type Position struct {

	// The quantity of the position.
	Quantity int32 `json:"quantity"`

	Product Product `json:"product"`

	// The total price of this position.
	Price float32 `json:"price"`

	// The total savings of this position.
	SavedPrice float32 `json:"savedPrice,omitempty"`
}